breaks out extensions
