+++ mapbook-iOS/AppDelegate.swift
@@ -95,92 +95,3 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         UIView.appearance(whenContainedInInstancesOf: [UIAlertController.self]).tintColor = UIColor.primaryBlue()
     }
 }
-
-extension UIColor {
-    
-    class func primaryBlue() -> UIColor {
-        return UIColor(red: 61.0/255.0, green: 81.0/255.0, blue: 180.0/255.0, alpha: 1)
-    }
-}
-
-extension Bundle {
-    
-    private static let agsBundle = AGSBundle()
-    
-    /// An end-user printable string representation of the ArcGIS Bundle version shipped with the app.
-    ///
-    /// For example, "2000"
-    
-    static var sdkBundleVersion: String {
-        return (agsBundle?.object(forInfoDictionaryKey: "CFBundleVersion") as? String) ?? "?"
-    }
-    
-    /// An end-user printable string representation of the ArcGIS Runtime SDK version shipped with the app.
-    ///
-    /// For example, "100.0.0"
-    
-    static var sdkVersion: String {
-        return (agsBundle?.object(forInfoDictionaryKey: "CFBundleShortVersionString") as? String) ?? "?"
-    }
-    
-    /// Builds an end-user printable string representation of the ArcGIS Bundle shipped with the app.
-    ///
-    /// For example, "ArcGIS Runtime SDK 100.0.0 (2000)"
-    
-    static var ArcGISSDKVersionString: String {
-        return "ArcGIS Runtime SDK \(sdkVersion) (\(sdkBundleVersion))"
-    }
-}
-
-extension Bundle {
-    
-    /// An end-user printable string representation of the app display name.
-    ///
-    /// For example, "Data Collection"
-    
-    static var appDisplayName: String {
-        return (main.object(forInfoDictionaryKey: "CFBundleDisplayName") as? String) ?? "?"
-    }
-    
-    /// An end-user printable string representation of the app version number.
-    ///
-    /// For example, "1.0"
-    
-    static var appVersion: String {
-        return (main.object(forInfoDictionaryKey: "CFBundleShortVersionString") as? String) ?? "?"
-    }
-    
-    /// An end-user printable string representation of the app bundle number.
-    ///
-    /// For example, "10"
-    
-    static var appBundleVersion: String {
-        return (main.object(forInfoDictionaryKey: "CFBundleVersion") as? String) ?? "?"
-    }
-    
-    /// Builds an end-user printable string representation of the app name and version.
-    ///
-    /// For example, "Data Collection 1.0 (10)"
-    
-    static var AppNameVersionString: String {
-        return "\(appDisplayName) \(appVersion) (\(appBundleVersion))"
-    }
-}
-
-extension AGSPortalInfo {
-    
-    var appPortalDescription: String? {
-        
-        var portalDescription: String = ""
-        if let organizationName = organizationName {
-            portalDescription = organizationName
-        }
-        if let portalName = portalName {
-            if portalDescription.count > 0 {
-                portalDescription += "\n"
-            }
-            portalDescription += portalName
-        }
-        return portalDescription.count > 0 ? portalDescription : nil
-    }
-}
