+++ mapbook-iOS/Extensions/AGSPortalInfo+Description.swift
@@ -1,0 +1,33 @@
+//// Copyright 2018 Gagandeep Singh
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import ArcGIS
+
+extension AGSPortalInfo {
+    
+    var appPortalDescription: String? {
+        
+        var portalDescription: String = ""
+        if let organizationName = organizationName {
+            portalDescription = organizationName
+        }
+        if let portalName = portalName {
+            if portalDescription.count > 0 {
+                portalDescription += "\n"
+            }
+            portalDescription += portalName
+        }
+        return portalDescription.count > 0 ? portalDescription : nil
+    }
+}
